refactored book room form component, removed context
