--- client/src/components/Form/Reservation/Context.js
@@ -1,4 +1,0 @@
-import React from 'react';
-
-const ThemeContext = React.createContext('light');
-export default ThemeContext;
